requirements.txt
aiohttp
beautifulsoup4
pydantic
motor
apscheduler
pandas
python-dotenv
pytest
fastapi[standard]